[IMP] mail : restrict access to mail.alias
Currently employees are allowed to create aliases
The configuration of a mail.alias is tricky and a badly configured one
could lead to issues.

Restrict the manual creation only to administrator users
Setting an alias through the configuration of a record (e.g. setting an alias
on a project), will be done with sudo.
Setting alias or deleting linked records should not change and no addition
priviledge should be required.

Some models like account.journal do not inherit mail.alias.mixing so manually
creating aliases as super user (we are not using the mixing as an alias should
not be created for every record, only under specific conditions).

Creation of a mail.alias is done in sudo if the user has create access rights
on the linked model.
This is required as an alias is created automatically through the mixin
(thanks to the inherits)

Task ID: 1919277