[REF] web, *: share common parts of main frontend layouts
* http_routing, portal, rating, survey, website, website_survey,
  website_slides_survey

Before this commit, the final base layout of website was a fully
overridden layout of the one in portal, which was somehow a duplicated
one of the login one in web, which... so lots of duplicated code.

This commit is a first step towards a better organization:

1) The web app defines a frontend layout (to include base frontend
   assets), with a base company logo as header.

2) The http_routing app modifies that layout in place to include
   the custom translation URL (this allows all modules to use the
   translation system).

3) The portal app modifies that layout in place to include the base
   header, footer, .... It also uses a primary extension of it for
   portal pages.

   The survey app simply uses the above layout instead of defining its
   own (by primary extension to include its own assets for its own
   pages)

   Same goes for the rating app and pages.

4) The website app modifies that layout in place to include the UI
   assets, to add website UI, ... This allows to create frontend apps
   which do not depend on website, with a non duplicated layout that
   will be automatically adapted if website is ever installed.

This work should be done even deeper as the layout parts of web /
http_routing is unclear (maybe all should be in web for this).

Many frontend but not website apps (not only survey / rating) could
probably also use this but this would be the topic of another task.

Part of https://github.com/odoo/odoo/pull/32181
task-1961045