[IMP] base: child's company_id should be same as parent company
purpose of the commit is When creating a new child system tries
to log a message on all other children and parents.
But since one of them is from a different company
the record rule triggers an error.

so if parent is set on contact then the company_id should be readonly and
it will automaitcally set the parent's company_id and
if parent's company_id changed then also change the company_id of all childs

Task-1934813