[IMP] point_of_sale: make 'PoS Orders' the default operation type
To improve the onboarding experience of the Point of Sale, a default operation type should be set for new Points of Sale as the setting is buried deep in the form and not immediately visible. If the default operation type does not suit the needs of the user, it can be changed later.

Related to task #1970513