[IMP] web: cache static resources forever
Today, some static resources are requested on every page load:
the menus, the translations and the qweb templates

After this commit, when loading the /web/ page, the user session will
contain the cache key to be used to fetch the resources. As the content
of the resource itself it used for the hash, it can be put in cache
forever on the browser.

The up side is that in general less requests are needed to load the web client
The down side is that the load time and number of queries of /web page
has increased to include what is needed to compute the hash of translations,
menus and qweb templates

Co-authored-by: Julien Mougenot <jum@odoo.com>