[IMP] web, website, *: modify templates to display relevant info
* = account

Multiple changes to enhance error displaying:
- errors messages are no longer the exception args but its name if the
exception is an except_orm. Else there is no message since the current
logic override exceptions of type Exception handled in ir_http.
- now on errors the traceback is only visible if in debug mode
- crashmanager.xml: javascript code removed, it is no
longer needed using WarningDialog.
- all except_orm exceptions are now displayed as a 400 except
AccessError wich is a 403
- WarningDialog and ErrorDialog now take the error as an argument
instead of in the options, this way its not passed to Dialog.

task-1894820