[IMP] crm: onchange of customer from quick create updates details in lead
Before commit:
When we create an opportunity manually from the Kanban view and associate
a customer to the opportunity, the "Initial Contact Information" doesn't
appear in the "Followup" tab on the lead so we have to reassign a partner
second time on the opportunity.

When we creating opportunity from lead and select option 'linked with existing
customer' Then in the 'followup' Section should't it fillup the address.

After commit:
When we create an opportunity manually from the Kanban view and associate
a customer to the opportunity, the "Initial Contact Information" will appear
in the "Followup" tab on lead.

When we creating opportunity from lead and select option 'linked with existing
customer' Then in the 'followup' Section should it fillup the address.

onchange fields are not available on quick create a view so the values do not
updated on record so added the invisible fields to make onchange works fabulously

Vlaues are not available on covert opperatunity data method so the values do
not appear on FollowUp tab so added the required data to make a partner
values visible in FollowUp tab.

task-1974504