[IMP] sale_timesheet: add AA orders and timesheets
Add the sales orders pointing to the Analytic Account set on the
project in the "Sales Order" smart button and the timesheets pointing
to the projects sharing that AA in the "Timesheets" smart button
on Project Overview.

The purpose is to improve the tracking of costs and revenues linked to
a project.

Task-1975022