[IMP] mail, web: to keep the follower preferences dropdown open
Currently, when check/unchecks the follower preferences, a dropdown
closes every time and it is really time-consuming. This is happening
because when click on any follower preferences checkbox, another click
fire on an active 'tab' from 'setLocalState' and that not make any sense
to click on 'tab' which is already active.

That's why made a condition proper which will not be going to click on
the active 'tab' if it is already active.

Currently, qunit is working fine regarding 'dropdown should be opened'
because there is a simple form which is not containing any 'tab'. But
when the form contains any 'tab' then a dropdown is not remaining open.

That's why added 'tab' inside the form to check that dropdown should be
opened in both cases (i.e. with 'tab' also).

task-1960834