[IMP] hr_contract: Remove social secretariat field
Social secretariat is a Belgian concept. It does not make
sense to have this in the standard contract app.

Will be reintroduced in l10n_be_hr_payroll (or internal)
if really needed.

Move `job_id` in the right group to easily add fields
on the right with view inheritance.