[IMP] mass_mailing: fixed buttons alignment issue for outlook
If you add multiple buttons in the single row. Outlook was showing them into
separate lines.

An outlook hack was added previously to fix the button background issue but when
we add multiple inline buttons in single row then buttons are displayed
in separate lines. To fix this issue we have wrapped these buttons inside the
table cells.