[IMP] mail: easy access to record from follower invitation mail
Previously, when someone is added as follower to the task, he/she is notified
via mail but, there was not any link to access that record easily.

So that, added a template that will show the mail along with the button,
onclick of that button follower can easily redirect that particular record.

task-1967234