[FIX] pos_sale: performance in sales reporting
Since v12.0, fields containing totals and taxes for pos lines are stored
in database.
This report was apparently not adapted when this change was introduced,
leaving a complex CTE to compute 'tax included' prices based on joins
with the account_tax table (and the relational table between pos lines
and taxes).

In database with a lot of pos order lines, this can make this SQL view
quite slow - this impacts performances in the sale report itself and,
most of all, in the form views of products and product templates which
use the sale.report model to compute the sales_count field on products.

This also seems to fix a problem that hadn't been reported until now
where the sale.report does not correctly compute taxes for pos lines.