New solution to keep the main_object untouched (we need it to not break existing behavior such as edit in backend)
This solution will actually give the possibility to set a different object to
be used for SEO.

THis won't break stable if the module website is not updated but the only way
to benefit from this is to update it.

Also, this could be used by other existing (or not yet) pages where the same
object might be used on different page, that way we can keep the main object
and have different SEO on those pages.